pyproject.toml
[project]
name = "lang-learn-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.136.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.14",
    "langchain-pinecone>=0.2.13",
    "langgraph-checkpoint-postgres>=3.0.5",
    "langsmith>=0.7.32",
    "openai>=2.32.0",
    "pinecone-client>=6.0.0",
    "psycopg2>=2.9.11",
    "pypdf>=6.10.2",
    "sqlmodel>=0.0.38",
]
